[Backend] Split pkg/plugins/handler.go into per-operation files
handler.go had grown to 2266 lines mixing install, update, manifest, order,
repositories, fields, listing, search, apply, persist, and conversion
concerns. Splitting it into focused files matches the existing per-operation
test file convention and makes targeted changes easier to review.
